reorganized a couple files
