+++ src/commands/Admin/Missy's Commands/dm.ts
@@ -1,7 +1,7 @@
 import { DMChannel } from 'discord.js';
 import { CommandStore, KlasaMessage, KlasaUser } from 'klasa';
 import MissyClient from '../../../lib/MissyClient';
-import MissyCommand from '../../../lib/structures/MissyCommand';
+import MissyCommand from '../../../lib/structures/base/MissyCommand';
 
 export default class extends MissyCommand {
