+++ src/commands/Admin/exec.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2017-2018 dirigeants. All rights reserved. MIT license.
 import { CommandStore, KlasaMessage } from 'klasa';
 import MissyClient from '../../lib/MissyClient';
-import MissyCommand from '../../lib/structures/MissyCommand';
+import MissyCommand from '../../lib/structures/base/MissyCommand';
 import { exec, codeBlock, scalarOrFirst } from '../../lib/util/util';
 
 export default class extends MissyCommand {
