+++ src/commands/Fun/Image/meme.ts
@@ -2,7 +2,7 @@ import assert from 'assert';
 import { Canvas } from 'canvas-constructor';
 import { CommandStore, KlasaMessage } from 'klasa';
 import MissyClient from '../../../lib/MissyClient';
-import MissyCommand from '../../../lib/structures/MissyCommand';
+import MissyCommand from '../../../lib/structures/base/MissyCommand';
 
 export default class extends MissyCommand {
