+++ src/commands/General/emoji.ts
@@ -1,5 +1,5 @@
 import { CommandStore, KlasaMessage } from 'klasa';
-import MissyCommand from '../../lib/structures/MissyCommand';
+import MissyCommand from '../../lib/structures/base/MissyCommand';
 import MissyClient from '../../lib/MissyClient';
 
 export default class extends MissyCommand {
