+++ src/commands/Scolding/eat.ts
@@ -1,7 +1,7 @@
 import { User } from 'discord.js';
 import { CommandStore, KlasaMessage } from 'klasa';
 import MissyClient from '../../lib/MissyClient';
-import MissyCommand from '../../lib/structures/MissyCommand';
+import MissyCommand from '../../lib/structures/base/MissyCommand';
 
 export default class extends MissyCommand {
