+++ src/commands/User Interaction/birthday.ts
@@ -1,7 +1,7 @@
 import assert from 'assert';
 import { CommandStore, KlasaMessage, KlasaUser } from 'klasa';
 import MissyClient from '../../lib/MissyClient';
-import MissyCommand from '../../lib/structures/MissyCommand';
+import MissyCommand from '../../lib/structures/base/MissyCommand';
 
 export default class extends MissyCommand {
